update directory ownership for ganeti 3
