+++ kitchen.yml
@@ -40,8 +40,6 @@ suites:
       ganeti:
         master-node: true
         drbd: true
-    provisioner:
-      multiple_converge: 3 # to fix directory ownership contention
   - name: drbd
     run_list:
       - recipe[ganeti-test]
@@ -50,8 +48,6 @@ suites:
       ganeti:
         master-node: true
         drbd: true
-    provisioner:
-      multiple_converge: 3 # to fix directory ownership contention
   - name: kvm
     run_list:
       - recipe[ganeti]
